requirements.txt
# MataVex Technologies - Core Node Dependencies

# Web & API Framework
Flask==3.1.3
flask-cors==6.0.2

# Database Connectivity
psycopg2-binary==2.9.11

# Security & Authentication
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
PyJWT==2.11.0

# Environment Management
python-dotenv==1.2.1

# Payment Integration (Optional/Extended)
razorpay==2.0.0

# Document & PDF Generation
reportlab==4.4.10
pillow==12.1.1

# Utilities
requests==2.32.5
pydantic==2.12.5
